+++ app/lib/fake/backend/fake_email_sender.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:mailer/mailer.dart';
+
 import '../../frontend/email_sender.dart';
 import '../../shared/email.dart';
 
@@ -9,9 +11,15 @@ FakeEmailSender get fakeEmailSender => emailSender as FakeEmailSender;
 
 class FakeEmailSender implements EmailSender {
   final sentMessages = <EmailMessage>[];
+  int failNextMessageCount = 0;
 
   @override
   Future<void> sendMessage(EmailMessage message) async {
+    message.verifyUuid();
+    if (failNextMessageCount > 0) {
+      failNextMessageCount--;
+      throw SmtpClientCommunicationException('fake network problem');
+    }
     sentMessages.add(message);
     // also trigger logging as fake pub server's integration test expect them
     await loggingEmailSender.sendMessage(message);